requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
matplotlib==3.7.2
Pillow==10.0.0
torch==2.0.1
torchvision==0.15.2
timm==0.9.7
transformers==4.30.0
jupyter==1.0.0
nbformat==5.9.2
nbconvert==7.8.0
accelerate==0.20.0 